// src/types/company.ts
export interface ICompanyInfoLinks {
	website: string;
	flickr: string;
	twitter: string;
	elon_twitter: string;
}

export interface ICompanyInfoHeadquarters {
	address: string;
	city: string;
	state: string;
}

export interface ICompanyInfo {
	headquarters: ICompanyInfoHeadquarters;
	links: ICompanyInfoLinks;
	name: string;
	founder: string;
	founded: number;
	employees: number;
	vehicles: number;
	launch_sites: number;
	test_sites: number;
	ceo: string;
	cto: string;
	coo: string;
	cto_propulsion: string;
	valuation: number;
	summary: string;
	id: string;
}